add dialog confirmation for city deletion in CitiesScreen

+++ app/src/main/java/com/arnedo/weatherapp/cities/view/CitiesScreen.kt
@@ -8,6 +8,9 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
@@ -18,6 +21,7 @@ import com.arnedo.weatherapp.cities.viewmodel.CityUiState
 import com.arnedo.weatherapp.cities.viewmodel.ICitiesViewModel
 import com.arnedo.weatherapp.common.entities.City
 import com.arnedo.weatherapp.common.model.getAllCityPreview
+import com.arnedo.weatherapp.ui.components.MyDialogInfo
 import com.arnedo.weatherapp.ui.components.MyProgressFullScreen
 import com.arnedo.weatherapp.ui.components.MyTextTitle
 import com.arnedo.weatherapp.ui.theme.WeatherAppTheme
@@ -30,6 +34,9 @@ fun CitiesView(
     modifier: Modifier,
     vm : ICitiesViewModel = koinViewModel<CitiesViewModel>()) {
     val uiState by vm.getUiState().collectAsState()
+    var openDialog by remember { mutableStateOf(false) }
+    var selectedCity by remember {mutableStateOf<City?>(null)}
+
     Box(modifier.fillMaxSize()) {
         Column(modifier.fillMaxSize(), horizontalAlignment = Alignment.CenterHorizontally) {
             MyTextTitle(R.string.cities_title)
@@ -39,10 +46,22 @@ fun CitiesView(
                     ItemCityView(
                         city = city,
                         onMap = { vm.showMap(city) },
-                        onRemove = {
-                            vm.deleteCity(city)
+                        onRemove = { city ->
+                            selectedCity = city
+                            openDialog = true
                         }
                         )
+                }
+            }
+            if(openDialog){
+                selectedCity?.let{ city ->
+                    MyDialogInfo(
+                        infoRes = R.string.dialog_msg_warning,
+                        titleRes = R.string.dialog_delete_title,
+                        confirmRes = R.string.dialog_delete_confirm){ isDeleted ->
+                            if(isDeleted) vm.deleteCity(city)
+                            openDialog = false
+                        }
                 }
             }
             MyProgressFullScreen(visible = uiState.inProgress)
